rough pass at breaking logspout.go into separate packages. also sped up using 'make dev'
