+++ modules.go
@@ -1,0 +1,6 @@
+package main
+
+import (
+	_ "github.com/gliderlabs/logspout/httpstream"
+	_ "github.com/gliderlabs/logspout/routesapi"
+)
